requirements.txt
asgiref==3.7.2
Django==4.2.10
django-js-asset==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
